+++ src/Math/Vector4.cpp
@@ -7,19 +7,19 @@ namespace umath
 	vector4::vector4()
 			: x(0), y(0),z(0),w(0)
 	{}
-	vector4::vector4(const vector3 Vec3, const double W)
+	vector4::vector4(const vector3 Vec3, const float W)
 		: 
 		x(Vec3[0]),
 		y(Vec3[1]),
 		z(Vec3[2]),
 		w(W)
 	{}
-	vector4::vector4(const double X, const double Y, const double Z, const double W)
+	vector4::vector4(const float X, const float Y, const float Z, const float W)
 		: x(X), y(Y), z(Z), w(W)
 	{}
 	vector4::~vector4(){}
 
-	double vector4::getLenght()
+	float vector4::getLenght()
 	{	
 		return sqrt(pow(x,2) + pow(y,2) + pow(z,2) + pow(w,2));
 	}
@@ -27,11 +27,11 @@ namespace umath
 	// Public
 	// Operators
 	
-	const double& vector4::operator [](const unsigned int& index) const
+	const float& vector4::operator [](const unsigned int& index) const
 	{
 		return (&x)[index];
 	}
-	double& vector4::operator [](const unsigned int& index)
+	float& vector4::operator [](const unsigned int& index)
 	{
 		return (&x)[index];
 	}
@@ -68,16 +68,16 @@ namespace umath
 		return LeftVal;
 	}
 
-	vector4 operator /(const vector4& LeftVal, const double& RightVal)
+	vector4 operator /(const vector4& LeftVal, const float& RightVal)
 	{
 		return vector4(LeftVal.x / RightVal,LeftVal.y / RightVal, LeftVal.z / RightVal, LeftVal.w / RightVal);
 	}
-	const vector4& operator /=(vector4& LeftVal, const double& RightVal)
+	const vector4& operator /=(vector4& LeftVal, const float& RightVal)
 	{
-		LeftVal.x /= (double)RightVal;
-		LeftVal.y /= (double)RightVal;
-		LeftVal.z /= (double)RightVal;
-		LeftVal.w /= (double)RightVal;
+		LeftVal.x /= (float)RightVal;
+		LeftVal.y /= (float)RightVal;
+		LeftVal.z /= (float)RightVal;
+		LeftVal.w /= (float)RightVal;
 		return LeftVal;
 	}
 
@@ -92,16 +92,16 @@ namespace umath
 		return !(LeftVal == RightVal);
 	}
 
-	vector4 operator *(const vector4& LeftVal, const double& RightVal)
+	vector4 operator *(const vector4& LeftVal, const float& RightVal)
 	{
 		return vector4(LeftVal.x * RightVal,LeftVal.y * RightVal, LeftVal.z * RightVal, LeftVal.w * RightVal);
 	}
-	vector4 operator *(const double& LeftVal, const vector4& RightVal)
+	vector4 operator *(const float& LeftVal, const vector4& RightVal)
 	{
 		return vector4(LeftVal * RightVal.x,LeftVal * RightVal.y, LeftVal * RightVal.z, LeftVal * RightVal.w);
 	}
 
-	const vector4& operator *=(vector4& LeftVal,const double& RightVal)
+	const vector4& operator *=(vector4& LeftVal,const float& RightVal)
 	{
 		LeftVal.x *= RightVal;
 		LeftVal.y *= RightVal;
